NextMove updates
Added interpretation function in GameMap to make it easy for the
players to know how much time is left.
Changed the instance handler - the player is now not allowed to move on
GameState creation, but only on the first GameState update. This is
because at GameState creation, the engine will wait for both players to
create their joysticks. After that step is complete, the 1 second timer
starts. So the players should not get to recalculate anything before
the timer starts.